+++ src/App.jsx
@@ -2,8 +2,7 @@ import { createBrowserRouter, RouterProvider } from "react-router-dom";
 import { CreateCampaign } from "./pages/campaign/createCampaign";
 import { CampaignManagementPage } from "./pages/campaign/CampaignManagementPage";
 import LogFormatPage from "./pages/format/format";
-import LogFilter from "./components/LogFilter";
-import SamplemainPage from "./pages/main/samplemain";
+import LogFilter from "./components/filter/LogFilter";
 import FilterManagementPage from "./pages/filter/filterManagementPage";
 
 export function App() {
