+++ src/pages/campaign/CampaignManagementPage.jsx
@@ -1,6 +1,6 @@
 import React from "react";
 import { Layout } from "../../components/Layout";
-import { CampaignTable } from "../../components/CampaginTable";
+import { CampaignTable } from "../../components/campaign/CampaignTable";
 
 export function CampaignManagementPage() {
   return (
