+++ src/pages/filter/filterManagementPage.jsx
@@ -1,5 +1,5 @@
 import React from "react";
-import FilterManagement from "../../components/FilterManagement";
+import FilterManagement from "../../components/filter/FilterManagement";
 
 const FilterManagementPage = () => {
   return (
